requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.2.0
matplotlib>=3.4.2
numpy>=1.20.3
opencv-python>=4.5.2
scikit-learn>=0.24.2
streamlit>=1.0.0
tqdm>=4.61.1